requirements.txt
pandas
numpy
sentence-transformers
faiss-cpu
fastapi
uvicorn
python-dotenv
ibm-watsonx-ai
streamlit
requests